SendingMoney splited into component
